Move load.py to audio/load.py
