+++ doc/index.rst
@@ -10,13 +10,12 @@ Welcome to opensoundscape's documentation!
 .. automodule:: opensoundscape.console
    :members:
 
-.. automodule:: opensoundscape.load
+.. automodule:: opensoundscape.audio.load
    :members:
 
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
-
 
 
 Indices and tables
