+++ pyproject.toml
@@ -18,6 +18,7 @@ include = ["opensoundscape/**/*.py"]
 python = "^3.6"
 docopt = "^0.6.2"
 librosa = "^0.7.0"
+ray = "^0.7.3"
 
 [tool.poetry.dev-dependencies]
 black = "=19.3b0"
